requirements.txt
# Core deep learning framework
torch>=1.9.0
torchvision>=0.10.0

# Numerical computing and data handling
numpy>=1.19.0
scipy>=1.7.0
pandas>=1.3.0

# Computer vision and image processing
opencv-python>=4.5.0
Pillow>=8.0.0

# Visualization and plotting
matplotlib>=3.3.0

# Progress bars and utilities
tqdm>=4.60.0

# Experiment tracking (optional but recommended)
wandb>=0.12.0

# Jupyter notebook support
jupyter>=1.0.0
ipykernel>=6.0.0